requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
exceptiongroup==1.2.2
fastapi==0.115.7
filelock==3.17.0
fonttools==4.55.4
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.15.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tifffile==2025.1.10
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
